Update from origin

+++ src/main/java/org/b3log/solo/processor/console/StaticSiteConsole.java
@@ -2,18 +2,12 @@
  * Solo - A small and beautiful blogging system written in Java.
  * Copyright (c) 2010-present, b3log.org
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * Solo is licensed under Mulan PSL v2.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *         http://license.coscl.org.cn/MulanPSL2
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ * See the Mulan PSL v2 for more details.
  */
 package org.b3log.solo.processor.console;
 
@@ -21,16 +15,15 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.time.DateFormatUtils;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.b3log.latke.Keys;
 import org.b3log.latke.Latkes;
-import org.b3log.latke.http.HttpMethod;
 import org.b3log.latke.http.RequestContext;
-import org.b3log.latke.http.annotation.Before;
-import org.b3log.latke.http.annotation.RequestProcessing;
-import org.b3log.latke.http.annotation.RequestProcessor;
 import org.b3log.latke.ioc.BeanManager;
 import org.b3log.latke.ioc.Inject;
-import org.b3log.latke.logging.Level;
-import org.b3log.latke.logging.Logger;
+import org.b3log.latke.ioc.Singleton;
 import org.b3log.latke.service.LangPropsService;
 import org.b3log.latke.util.CollectionUtils;
 import org.b3log.latke.util.Strings;
@@ -43,6 +36,7 @@ import org.json.JSONObject;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.OutputStream;
+import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 
@@ -50,17 +44,16 @@ import java.util.List;
  * Static site console request processing. HTML 静态站点生成 https://github.com/88250/solo/issues/19
  *
  * @author <a href="http://88250.b3log.org">Liang Ding</a>
- * @version 1.0.0.0, Jan 13, 2020
+ * @version 2.0.1.2, Mar 29, 2020
  * @since 3.9.0
  */
-@RequestProcessor
-@Before(ConsoleAdminAuthAdvice.class)
+@Singleton
 public class StaticSiteConsole {
 
     /**
      * Logger.
      */
-    private static final Logger LOGGER = Logger.getLogger(StaticSiteConsole.class);
+    private static final Logger LOGGER = LogManager.getLogger(StaticSiteConsole.class);
 
     /**
      * Language service.
@@ -69,11 +62,20 @@ public class StaticSiteConsole {
     private LangPropsService langPropsService;
 
     /**
+     * Mock request scheme.
+     */
+    private static String scheme;
+
+    /**
+     * Mock request host.
+     */
+    private static String host;
+
+    /**
      * Generates static site.
      *
      * @param context the specified request context
      */
-    @RequestProcessing(value = "/console/staticsite", method = HttpMethod.PUT)
     public synchronized void genSite(final RequestContext context) {
         try {
             final JSONObject requestJSONObject = context.requestJSON();
@@ -85,13 +87,22 @@ public class StaticSiteConsole {
                 return;
             }
 
+            if (Latkes.isInJar()) {
+                context.renderJSON(-1);
+                context.renderMsg("Do not support this feature while running in Jar");
+
+                return;
+            }
+
             FileUtils.deleteDirectory(new File(staticSitePath));
             FileUtils.forceMkdir(new File(staticSitePath));
 
-            // 切换至静态站点生成模式
-            Latkes.setServerScheme("https");
-            Latkes.setServerHost(url);
-            Latkes.setServerPort("");
+            final URL u = new URL(url);
+            scheme = u.getProtocol();
+            host = u.getHost();
+            if (-1 != u.getPort()) {
+                host += ":" + u.getPort();
+            }
             Solos.GEN_STATIC_SITE = true;
 
             genURI("/tags.html");
@@ -115,10 +126,6 @@ public class StaticSiteConsole {
             genFile("robots.txt");
             genFile("CHANGE_LOGS.md");
 
-            // 恢复之前的动态运行模式
-            Latkes.setServerScheme("http");
-            Latkes.setServerHost("localhost");
-            Latkes.setServerPort("8080");
             Solos.GEN_STATIC_SITE = false;
 
             LOGGER.log(Level.INFO, "Static site generated [dir=" + staticSitePath + "]");
@@ -136,29 +143,44 @@ public class StaticSiteConsole {
     }
 
     /**
-     * Name of generate directory.
+     * Source root directory path.
      */
-    private static final String STATIC_SITE = "static-site";
+    private static String rootPath;
 
     /**
      * Path of generate directory.
      */
-    private static final String staticSitePath = StaticSiteConsole.class.getResource("/" + STATIC_SITE).getPath();
+    private static String staticSitePath;
 
-    /**
-     * Source directory path.
-     */
-    private static final String sourcePath = StaticSiteConsole.class.getResource("/").getPath();
 
-    private static void genCategories() throws Exception {
+    static {
+        if (!Latkes.isInJar()) {
+            rootPath = StaticSiteConsole.class.getResource("/repository.json").getPath();
+            rootPath = StringUtils.substringBeforeLast(rootPath, "/repository.json");
+            staticSitePath = rootPath + "/static-site";
+        } else {
+            LOGGER.log(Level.INFO, "Do not support export static site when running in jar");
+        }
+    }
+
+    private static void genCategories() {
         final BeanManager beanManager = BeanManager.getInstance();
         final CategoryQueryService categoryQueryService = beanManager.getReference(CategoryQueryService.class);
         final List<JSONObject> categories = categoryQueryService.getMostTagCategory(Integer.MAX_VALUE);
+        final OptionQueryService optionQueryService = beanManager.getReference(OptionQueryService.class);
+        final JSONObject preference = optionQueryService.getPreference();
+        final int pageSize = preference.getInt(Option.ID_C_ARTICLE_LIST_DISPLAY_COUNT);
 
         categories.parallelStream().forEach(category -> {
             final String categoryURI = category.optString(Category.CATEGORY_URI);
             try {
-                genPage("/category/" + categoryURI);
+                final int articleCount = categoryQueryService.getPublishedArticleCount(category.optString(Keys.OBJECT_ID));
+                final int pageCount = (int) Math.ceil((double) articleCount / pageSize);
+                int count = 0;
+                while (count++ < pageCount) {
+                    genPage("/category/" + categoryURI);
+                    genPage("/category/" + categoryURI + "?p=" + count);
+                }
             } catch (final Exception e) {
                 LOGGER.log(Level.ERROR, "Generates a category [uri=" + categoryURI + "] failed", e);
             }
@@ -169,12 +191,21 @@ public class StaticSiteConsole {
         final BeanManager beanManager = BeanManager.getInstance();
         final ArchiveDateQueryService archiveDateQueryService = beanManager.getReference(ArchiveDateQueryService.class);
         final List<JSONObject> archiveDates = archiveDateQueryService.getArchiveDates();
+        final OptionQueryService optionQueryService = beanManager.getReference(OptionQueryService.class);
+        final JSONObject preference = optionQueryService.getPreference();
+        final int pageSize = preference.getInt(Option.ID_C_ARTICLE_LIST_DISPLAY_COUNT);
 
         archiveDates.parallelStream().forEach(archiveDate -> {
             final long time = archiveDate.getLong(ArchiveDate.ARCHIVE_TIME);
             final String dateString = DateFormatUtils.format(time, "yyyy/MM");
             try {
-                genPage("/archives/" + dateString);
+                final int articleCount = archiveDateQueryService.getArchiveDatePublishedArticleCount(archiveDate.optString(Keys.OBJECT_ID));
+                final int pageCount = (int) Math.ceil((double) articleCount / pageSize);
+                int count = 0;
+                while (count++ < pageCount) {
+                    genPage("/archives/" + dateString);
+                    genPage("/archives/" + dateString + "?p=" + count);
+                }
             } catch (final Exception e) {
                 LOGGER.log(Level.ERROR, "Generates an archive [date=" + archiveDate + "] failed", e);
             }
@@ -185,11 +216,20 @@ public class StaticSiteConsole {
         final BeanManager beanManager = BeanManager.getInstance();
         final TagQueryService tagQueryService = beanManager.getReference(TagQueryService.class);
         final List<JSONObject> tags = tagQueryService.getTags();
+        final OptionQueryService optionQueryService = beanManager.getReference(OptionQueryService.class);
+        final JSONObject preference = optionQueryService.getPreference();
+        final int pageSize = preference.getInt(Option.ID_C_ARTICLE_LIST_DISPLAY_COUNT);
 
         tags.parallelStream().forEach(tag -> {
             final String tagTitle = tag.optString(Tag.TAG_TITLE);
             try {
-                genPage("/tags/" + tagTitle);
+                final int articleCount = tagQueryService.getArticleCount(tag.optString(Keys.OBJECT_ID));
+                final int pageCount = (int) Math.ceil((double) articleCount / pageSize);
+                int count = 0;
+                while (count++ < pageCount) {
+                    genPage("/tags/" + tagTitle);
+                    genPage("/tags/" + tagTitle + "?p=" + count);
+                }
             } catch (final Exception e) {
                 LOGGER.log(Level.ERROR, "Generates a tag [title=" + tagTitle + "] failed", e);
             }
@@ -235,7 +275,7 @@ public class StaticSiteConsole {
         filePath = StringUtils.replace(filePath, "=", "/");
         FileUtils.forceMkdir(new File(staticSitePath + filePath));
         final OutputStream outputStream = new FileOutputStream(staticSitePath + filePath + "/index.html");
-        String html = Mocks.mockRequest(uri);
+        String html = Mocks.mockRequest(uri, scheme, host);
         IOUtils.write(html, outputStream, StandardCharsets.UTF_8);
         outputStream.close();
         LOGGER.log(Level.INFO, "Generated a page [" + uri + "]");
@@ -244,7 +284,7 @@ public class StaticSiteConsole {
     private static void genURI(final String uri) throws Exception {
         FileUtils.forceMkdirParent(new File(staticSitePath + uri));
         final OutputStream outputStream = new FileOutputStream(staticSitePath + uri);
-        String html = Mocks.mockRequest(uri);
+        String html = Mocks.mockRequest(uri, scheme, host);
         IOUtils.write(html, outputStream, StandardCharsets.UTF_8);
         outputStream.close();
         LOGGER.log(Level.INFO, "Generated a file [" + uri + "]");
@@ -253,13 +293,13 @@ public class StaticSiteConsole {
     private static void genArticle(final String permalink) throws Exception {
         if (!StringUtils.endsWithIgnoreCase(permalink, ".html") && !StringUtils.endsWithIgnoreCase(permalink, ".htm")) {
             FileUtils.forceMkdir(new File(staticSitePath + permalink));
-            final String html = Mocks.mockRequest(permalink);
+            final String html = Mocks.mockRequest(permalink, scheme, host);
             final OutputStream outputStream = new FileOutputStream(staticSitePath + permalink + "/index.html");
             IOUtils.write(html, outputStream, StandardCharsets.UTF_8);
             outputStream.close();
         } else {
             FileUtils.forceMkdirParent(new File(staticSitePath + permalink));
-            final String html = Mocks.mockRequest(permalink);
+            final String html = Mocks.mockRequest(permalink, scheme, host);
             final OutputStream outputStream = new FileOutputStream(staticSitePath + permalink);
             IOUtils.write(html, outputStream, StandardCharsets.UTF_8);
             outputStream.close();
@@ -270,28 +310,28 @@ public class StaticSiteConsole {
     private static void genSkins() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/skins"));
         FileUtils.forceMkdir(new File(staticSitePath + "/skins"));
-        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/skins").toURI()), new File(staticSitePath + "/skins"));
+        FileUtils.copyDirectory(new File(rootPath + "/skins"), new File(staticSitePath + "/skins"));
         LOGGER.log(Level.INFO, "Generated skins");
     }
 
     private static void genJS() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/js"));
         FileUtils.forceMkdir(new File(staticSitePath + "/js"));
-        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/js").toURI()), new File(staticSitePath + "/js"));
+        FileUtils.copyDirectory(new File(rootPath + "/js"), new File(staticSitePath + "/js"));
         LOGGER.log(Level.INFO, "Generated js");
     }
 
     private static void genImages() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/images"));
         FileUtils.forceMkdir(new File(staticSitePath + "/images"));
-        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/images").toURI()), new File(staticSitePath + "/images"));
+        FileUtils.copyDirectory(new File(rootPath + "/images"), new File(staticSitePath + "/images"));
         LOGGER.log(Level.INFO, "Generated images");
     }
 
     private static void genPlugins() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/plugins"));
         FileUtils.forceMkdir(new File(staticSitePath + "/plugins"));
-        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/plugins").toURI()), new File(staticSitePath + "/plugins"));
+        FileUtils.copyDirectory(new File(rootPath + "/plugins"), new File(staticSitePath + "/plugins"));
         genURI("/plugins/kanbanniang/assets/model.json");
         LOGGER.log(Level.INFO, "Generated plugins");
 
@@ -299,8 +339,7 @@ public class StaticSiteConsole {
 
     private static void genFile(final String file) throws Exception {
         FileUtils.forceMkdirParent(new File(staticSitePath + "/" + file));
-        final String staticSitePath = StaticSiteConsole.class.getResource("/" + STATIC_SITE).toURI().getPath();
-        FileUtils.copyFile(new File(sourcePath + "/" + file), new File(staticSitePath + "/" + file));
+        FileUtils.copyFile(new File(rootPath + "/" + file), new File(staticSitePath + "/" + file));
         LOGGER.log(Level.INFO, "Generated a file [" + file + "]");
     }
